requirements.txt
deepseek==1.0.0
playwright==1.35.0
numpy==1.24.3
pandas==2.0.2
pyautogui==0.9.54
pywinctl==0.2.0
fastapi==0.99.1
uvicorn==0.22.0
python-dotenv==1.0.0
pywin32==306